dtu_mlops_69/frontend-backend/requirements_frontend.txt
torch==2.2.2
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
transformers==4.48.1
pillow
numpy==1.23.5
torchvision==0.17.2
google-cloud-run==0.10.14
streamlit==1.41.1